begin migrating some extensions stuff to extensions module
